requirements.txt
asgiref==3.7.2
build==1.0.3
click==8.1.7
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
env==0.1.0
Markdown==3.4.4
packaging==23.1
Pillow==10.0.1
pip-tools==7.3.0
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
